[tests/e2e] move helpers into a new sub-mod
