+++ tests/src/e2e.rs
@@ -12,6 +12,7 @@
 //! `ANOMA_E2E_KEEP_TEMP=true`.
 
 pub mod gossip_tests;
+pub mod helpers;
 pub mod ledger_tests;
 pub mod setup;
 pub mod wallet_tests;
