+++ tests/src/e2e/gossip_tests.rs
@@ -16,6 +16,7 @@ use color_eyre::eyre::Result;
 use serde_json::json;
 use setup::constants::*;
 
+use crate::e2e::helpers::find_address;
 use crate::e2e::setup::{self, Bin, Who};
 use crate::{run, run_as};
 
@@ -75,12 +76,12 @@ fn match_intents() -> Result<()> {
     let intent_b_path_input = test.base_dir.path().join("intent.B.data");
     let intent_c_path_input = test.base_dir.path().join("intent.C.data");
 
-    let albert = setup::find_address(&test, ALBERT)?;
-    let bertha = setup::find_address(&test, BERTHA)?;
-    let christel = setup::find_address(&test, CHRISTEL)?;
-    let xan = setup::find_address(&test, XAN)?;
-    let btc = setup::find_address(&test, BTC)?;
-    let eth = setup::find_address(&test, ETH)?;
+    let albert = find_address(&test, ALBERT)?;
+    let bertha = find_address(&test, BERTHA)?;
+    let christel = find_address(&test, CHRISTEL)?;
+    let xan = find_address(&test, XAN)?;
+    let btc = find_address(&test, BTC)?;
+    let eth = find_address(&test, ETH)?;
     let intent_a_json = json!([
         {
             "key": bertha,
